pyproject.toml
[project]
name = "pushup-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.2"
dependencies = [
    "aiogram==3.13.0",
    "python-dotenv==1.2.2",
    "ruff==0.15.11",
]
[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle (ошибки)
    "F",   # pyflakes (импорты, undefined)
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # bugbear (очень полезно)
    "C4",  # comprehensions
    "SIM", # упрощение кода
    "ASYNC", # async-антипаттерны 🔥
]

ignore = [
    "E501",  # длина строки (контролируем сами)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
